+++ packages/terra-form-select/tests/jest/ClearOption.test.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl, mountWithIntl } from '@cerner/terra-enzyme-intl';
 import ClearOption from '../../src/shared/_ClearOption';
 
 describe('ClearOption', () => {
   it('should render a default ClearOption', () => {
-    const wrapper = shallowWithIntl(<ClearOption value="" display="Clear" />);
+    const wrapper = shallowWithIntl(<ClearOption value="" display="Clear" />).dive();
     expect(wrapper).toMatchSnapshot();
   });