+++ packages/terra-form-select/tests/jest/MultiSelectField.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import MultiSelectField from '../../src/MultiSelectField';
 
 it('should render MultiSelectField component with max height', () => {
@@ -14,6 +14,6 @@ it('should render MultiSelectField component with max height', () => {
     </MultiSelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });