+++ packages/terra-form-select/tests/jest/OptGroup.test.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 
 /* eslint-disable import/no-extraneous-dependencies */
-import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl, mountWithIntl } from '@cerner/terra-enzyme-intl';
 import OptGroup from '../../src/shared/_OptGroup';
 
 describe('OptGroup', () => {
   it('should render a default Option', () => {
-    const wrapper = shallowWithIntl(<OptGroup label="label" />);
+    const wrapper = shallowWithIntl(<OptGroup label="label" />).dive();
     expect(wrapper).toMatchSnapshot();
   });