+++ packages/terra-form-select/tests/jest/SelectField.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import SelectField from '../../src/SelectField';
 
 it('should render a default SelectField component', () => {
@@ -14,7 +14,7 @@ it('should render a default SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a disabled SelectField component', () => {
@@ -28,7 +28,7 @@ it('should render a disabled SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a disabled SelectField component via selectAttrs', () => {
@@ -42,7 +42,7 @@ it('should render a disabled SelectField component via selectAttrs', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a default SelectField component with attributes spread onto the select', () => {
@@ -56,7 +56,7 @@ it('should render a default SelectField component with attributes spread onto th
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a multiple SelectField component that limits the selection count', () => {
@@ -70,7 +70,7 @@ it('should render a multiple SelectField component that limits the selection cou
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a required SelectField component', () => {
@@ -84,7 +84,7 @@ it('should render a required SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render a max height SelectField component', () => {
@@ -98,7 +98,7 @@ it('should render a max height SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });
 
 it('should render an incomplete SelectField component', () => {
@@ -112,5 +112,5 @@ it('should render an incomplete SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallowWithIntl(select).dive().dive()).toMatchSnapshot();
+  expect(shallowWithIntl(select).dive().dive().dive()).toMatchSnapshot();
 });