+++ packages/terra-form-select/tests/jest/native-select/NativeSelectField.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import NativeSelectField from '../../../src/native-select/NativeSelectField';
 
 describe('NativeSelectField', () => {