+++ packages/terra-show-hide/tests/jest/ShowHide.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { mountWithIntl } from 'terra-enzyme-intl';
+import { mountWithIntl } from '@cerner/terra-enzyme-intl';
 import ShowHide from '../../src/ShowHide';
 
 describe('ShowHide', () => {