+++ packages/terra-i18n/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated Jest Snapshots
+
 ## 4.38.1 - (January 4, 2022)
 
 * Changed